fix: remove backend
