+++ open_gpt/cli/commands/deploy.py
@@ -31,12 +31,6 @@ class DeployCommand(Command):
             default='balanced',
         ),
         option(
-            'backend',
-            None,
-            'The backend of the model.',
-            default='hf',
-        ),
-        option(
             "replicas", "r", "The number of replicas to serve.", flag=False, default=1
         ),
         option(
@@ -76,7 +70,6 @@ class DeployCommand(Command):
                         'precision': self.option('precision'),
                         'adapter_name_or_path': self.option('adapter_name_or_path'),
                         'device_map': self.option('device_map'),
-                        'backend': self.option('backend'),
                     },
                     replicas=self.option('replicas'),
                     instance_type=self.option('instance_type'),
